adds swagger docs
