+++ gateway/docs/swagger/evm-routes.yml
@@ -1,0 +1,76 @@
+paths: 
+  /evm/nonce:
+    post:
+      summary: 'Get the nonce for the provided private key'
+      operationId: 'nonce'
+      consumes:
+        - 'application/json'
+      produces:
+        - 'application/json'
+      parameters:
+        - in: 'body'
+          name: 'body'
+          required: true
+          schema:
+            $ref: '#/definitions/NonceRequest'
+      responses:
+        '200':
+          schema:
+            $ref: '#/definitions/NonceResponse'
+
+  /evm/allowances:
+    post:
+      summary: 'Get the ERC20 allowances for a spender on a given private key'
+      operationId: 'allowances'
+      consumes:
+        - 'application/json'
+      produces:
+        - 'application/json'
+      parameters:
+        - in: 'body'
+          name: 'body'
+          required: true
+          schema:
+            $ref: '#/definitions/AllowancesRequest'
+      responses:
+        '200':
+          schema:
+            $ref: '#/definitions/AllowancesResponse'
+/evm/approve:
+    post:
+      summary: 'Create an ERC20 approval for a spender on an private key'
+      operationId: 'approve'
+      consumes:
+        - 'application/json'
+      produces:
+        - 'application/json'
+      parameters:
+        - in: 'body'
+          name: 'body'
+          required: true
+          schema:
+            $ref: '#/definitions/ApproveRequest'
+      responses:
+        '200':
+          schema:
+            $ref: '#/definitions/ApproveResponse'
+
+  /evm/cancel:
+    post:
+      summary: 'Cancel transaction'
+      operationId: 'cancel'
+      consumes:
+        - 'application/json'
+      produces:
+        - 'application/json'
+      parameters:
+        - in: 'body'
+          name: 'body'
+          required: true
+          schema:
+            $ref: '#/definitions/CancelRequest'
+      responses:
+        '200':
+          schema:
+            $ref: '#/definitions/CancelResponse'
+
