+++ gateway/src/services/config-manager-v2.ts
@@ -361,6 +361,7 @@ export class ConfigManagerV2 {
 
   loadConfigRoot(configRootPath: string) {
     // Load the config root file.
+    console.log(configRootPath);
     const configRootFullPath: string = fs.realpathSync(configRootPath);
     const configRootTemplateFullPath: string = path.join(
       ConfigTemplatesDir,
